refactor: Move domain classes to a new model package and remove unused classes
