+++ src/main/java/engineer/mkitsoukou/ntango/domain/model/Interval.java
@@ -1,4 +1,4 @@
-package engineer.mkitsoukou.ntango.domain;
+package engineer.mkitsoukou.ntango.domain.model;
 
 public record Interval(
     int start,
