+++ src/main/java/engineer/mkitsoukou/ntango/domain/model/ScheduleResult.java
@@ -1,4 +1,4 @@
-package engineer.mkitsoukou.ntango.domain;
+package engineer.mkitsoukou.ntango.domain.model;
 
 public record ScheduleResult(
     String taskId,
